+++ src/main/java/com/example/ejerciciohibernateandres/dao/ExpertoDAO.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface ExpertoDAO {
 
-    List<Experto> encontrarConFiltros(String nombre, Integer etiqueta, String modalidad, String estado);
+    List<Experto> encontrarConFiltros(String nombre, Long etiqueta, String modalidad, String estado);
 
 }
